stereo mix buffers
- Mixer now filling a buffer per output channel
- added AudioDevice::ConvertFormat()
- added Voice::GetSample()
- added Voice::GetSamplesRemaining()
- added Voice::AdvancePlayHead()
- added Sound::GetSample() including format conversion

+++ src/YoAudio.cpp
@@ -26,7 +26,7 @@ bool YOA_StopVoice(const int id, const float fadeOut)
 {
 	AudioThread* inst = AudioThread::GetInstance();
 	if (inst != nullptr) {
-		return inst->mMixer->StopVoice(id, fadeOut);
+		return inst->mMixer->StopVoice(uint16_t(id), fadeOut);
 	}
 	return false;
 }
